requirements.txt
python-telegram-bot==20.7
apscheduler==3.10.4
pytz==2024.1
openai==1.30.0
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.29.0
google-api-python-client==2.127.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
tenacity==8.3.0
aiosqlite==0.20.0
matplotlib==3.8.4
